Move the HIH6130 files into the ethoscope.hardware.input package, and the tests into ethoscope.tests.unittests
