+++ src/ethoscope/hardware/input/HIH6130.py
