+++ src/ethoscope/hardware/input/easyi2c.py
